Move Toils into their own subfolder and adjust namespace
